Break out functions into separate modules
This file is starting to become a bit unruly. I'm breaking these out to
make the code more manageable and to eventually improve our testing
story.
