+++ src/constants.js
@@ -1,0 +1,6 @@
+export default {
+  above: 'above',
+  inside: 'inside',
+  below: 'below',
+  invisible: 'invisible',
+};
